requirements.txt
# Vehicle Detection System - Core Dependencies

# Core computer vision and ML
opencv-python>=4.6.0
numpy>=1.21.0
PyYAML>=6.0

# GUI framework (optional but recommended)
PySide6>=6.4.0

# OpenVINO for inference (optional - fallback to mock if not available)
openvino>=2022.3.0

# Data handling
pandas>=1.3.0
requests>=2.25.0

# Utilities
Pillow>=8.3.0
matplotlib>=3.5.0
